# Conflicts:
#	TestModels/SimpleTypes/SimpleBoolean/codegen-patches/rust/dafny-4.5.0.patch
#	TestModels/SimpleTypes/SimpleBoolean/runtimes/rust/dafny_impl/src/implementation_from_dafny.rs
#	TestModels/SimpleTypes/SimpleBoolean/runtimes/rust/src/client.rs
#	TestModels/SimpleTypes/SimpleBoolean/runtimes/rust/src/conversions/get_boolean.rs
#	TestModels/SimpleTypes/SimpleBoolean/runtimes/rust/src/implementation_from_dafny.rs
#	TestModels/SimpleTypes/SimpleTimestamp/runtimes/rust/dafny_impl/src/implementation_from_dafny.rs
#	TestModels/dafny-dependencies/StandardLibrary/codegen-patches/rust/dafny-4.5.0.patch
#	codegen/smithy-dafny-codegen/src/main/java/software/amazon/polymorph/CodegenEngine.java

+++ codegen/smithy-dafny-codegen/src/main/java/software/amazon/polymorph/CodegenEngine.java
@@ -8,6 +8,7 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Streams;
 
 import java.io.File;
+import com.squareup.javapoet.ClassName;
 import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.nio.file.Files;
@@ -29,6 +30,7 @@ import org.slf4j.LoggerFactory;
 import software.amazon.polymorph.smithydafny.DafnyApiCodegen;
 import software.amazon.polymorph.smithydafny.DafnyNameResolver;
 import software.amazon.polymorph.smithydafny.DafnyVersion;
+import software.amazon.polymorph.smithydotnet.AwsSdkDotNetNameResolver;
 import software.amazon.polymorph.smithydotnet.AwsSdkShimCodegen;
 import software.amazon.polymorph.smithydotnet.AwsSdkTypeConversionCodegen;
 import software.amazon.polymorph.smithydotnet.DotNetNameResolver;
@@ -44,6 +46,7 @@ import software.amazon.polymorph.smithyjava.generator.awssdk.v1.JavaAwsSdkV1;
 import software.amazon.polymorph.smithyjava.generator.awssdk.v2.JavaAwsSdkV2;
 import software.amazon.polymorph.smithyjava.generator.library.JavaLibrary;
 import software.amazon.polymorph.smithyjava.generator.library.TestJavaLibrary;
+import software.amazon.polymorph.smithyjava.nameresolver.AwsSdkNativeV2;
 import software.amazon.polymorph.traits.LocalServiceTrait;
 import software.amazon.polymorph.utils.DafnyNameResolverHelpers;
 import software.amazon.polymorph.utils.IOUtils;
@@ -153,6 +156,8 @@ public class CodegenEngine {
       System.exit(1);
     }
 
+    propertiesFile.ifPresent(this::generateProjectPropertiesFile);
+
     for (final TargetLanguage lang : targetLangOutputDirs.keySet()) {
       final Path outputDir = targetLangOutputDirs
         .get(lang)
@@ -172,8 +177,6 @@ public class CodegenEngine {
         );
       }
     }
-
-    propertiesFile.ifPresent(this::generateProjectPropertiesFile);
   }
 
   private void generateProjectPropertiesFile(final Path outputPath)
@@ -310,7 +313,7 @@ public class CodegenEngine {
    * Generate a skeletal implementation of the local service operations,
    * with `expect false` statements to ensure tests will initially fail.
    */
-  public void generateDafnySkeleton(Path outputDir) {
+  private void generateDafnySkeleton(Path outputDir) {
     final DafnyApiCodegen dafnyApiCodegen = new DafnyApiCodegen(
       model,
       serviceShape,
@@ -340,8 +343,16 @@ public class CodegenEngine {
     LOGGER.info("Formatting Java code in {}", outputDir);
     runCommand(
       outputDir,
+      "npm",
+      "i",
+      "--no-save",
+      "prettier@3",
+      "prettier-plugin-java@2.5"
+    );
+    runCommand(
+      outputDir,
       "npx",
-      "prettier",
+      "prettier@3",
       "--plugin=prettier-plugin-java",
       outputDir.toString(),
       "--write"
@@ -418,8 +429,12 @@ public class CodegenEngine {
     final String sdkID = awsSdkStyle
       ? serviceShape.expectTrait(ServiceTrait.class).getSdkId()
       : serviceShape.expectTrait(LocalServiceTrait.class).getSdkId();
+    final String serviceName = sdkID.toLowerCase();
     final String namespace = serviceShape.getId().getNamespace();
     final String namespaceDir = namespace.replace(".", "/");
+    final String dafnyNamespace =
+      DafnyNameResolverHelpers.packageNameForNamespace(namespace);
+    final String dafnyNamespaceDir = dafnyNamespace.replace(".", "/");
 
     final String gradleGroup = namespace;
     // TODO: This should be @title, but we have to actually add that to all services first
@@ -429,8 +444,12 @@ public class CodegenEngine {
     parameters.put("dafnyVersion", dafnyVersion.unparse());
     parameters.put("service", service);
     parameters.put("sdkID", sdkID);
+    parameters.put("serviceName", serviceName);
     parameters.put("namespace", namespace);
     parameters.put("namespaceDir", namespaceDir);
+    parameters.put("dafnyNamespace", dafnyNamespace);
+    parameters.put("dafnyNamespaceDir", dafnyNamespaceDir);
+
     parameters.put("gradleGroup", gradleGroup);
     parameters.put("gradleDescription", gradleDescription);
 
@@ -441,7 +460,21 @@ public class CodegenEngine {
           parameters
         );
       }
-      // TODO generate sdk constructor
+      if (generationAspects.contains(GenerationAspect.CLIENT_CONSTRUCTORS)) {
+        final ClassName clientClassName =
+          AwsSdkNativeV2.classNameForServiceClient(serviceShape);
+        parameters.put("sdkClientNamespace", clientClassName.packageName());
+        parameters.put("sdkClientName", clientClassName.simpleName());
+
+        writeTemplatedFile(
+          "runtimes/java/src/main/java/$dafnyNamespaceDir;L/__default.java",
+          parameters
+        );
+        writeTemplatedFile(
+          "runtimes/java/src/main/java/$dafnyNamespaceDir;L/types/__default.java",
+          parameters
+        );
+      }
     } else {
       final String serviceConfig = serviceShape
         .expectTrait(LocalServiceTrait.class)
@@ -548,16 +581,18 @@ public class CodegenEngine {
   }
 
   private void netOtherGeneratedAspects() {
-    final DotNetNameResolver resolver = new DotNetNameResolver(
-      model,
-      serviceShape
-    );
+    final DotNetNameResolver resolver = awsSdkStyle
+      ? new AwsSdkDotNetNameResolver(model, serviceShape)
+      : new DotNetNameResolver(model, serviceShape);
     final String service = serviceShape.getId().getName();
     final String sdkID = awsSdkStyle
       ? serviceShape.expectTrait(ServiceTrait.class).getSdkId()
       : serviceShape.expectTrait(LocalServiceTrait.class).getSdkId();
+    final String serviceName = resolver.getServiceName();
     final String namespace = serviceShape.getId().getNamespace();
-    final String dotnetNamespace = resolver.namespaceForService();
+    final String dotnetNamespace = resolver.namespaceForShapeId(
+      serviceShape.getId()
+    );
     final String dafnyNamespace =
       DafnyNameResolverHelpers.packageNameForNamespace(namespace);
     final String namespaceDir = namespace.replace(".", "/");
@@ -566,12 +601,17 @@ public class CodegenEngine {
     parameters.put("dafnyVersion", dafnyVersion.unparse());
     parameters.put("service", service);
     parameters.put("sdkID", sdkID);
+    parameters.put("serviceName", serviceName);
     parameters.put("namespace", dotnetNamespace);
     parameters.put("dafnyNamespace", dafnyNamespace);
     parameters.put("namespaceDir", namespaceDir);
     parameters.put("stdLibPath", standardLibraryPath().toString());
 
     if (awsSdkStyle) {
+      final String clientClassName =
+        ((AwsSdkDotNetNameResolver) resolver).implForServiceClient();
+      parameters.put("sdkClientName", clientClassName);
+
       if (generationAspects.contains(GenerationAspect.PROJECT_FILES)) {
         writeTemplatedFile("runtimes/net/$forSDK;L$sdkID;L.csproj", parameters);
       }
@@ -628,17 +668,23 @@ public class CodegenEngine {
     // Be sure to NOT delete src/implementation_from_dafny.rs though,
     // by temporarily moving it out of src/
     Path outputSrcDir = outputDir.resolve("src");
-    Path implementationFromDafnyPath = outputSrcDir.resolve("implementation_from_dafny.rs");
-      try {
-        Path tmpPath = outputDir.resolve("implementation_from_dafny.rs");
+    Path implementationFromDafnyPath = outputSrcDir.resolve(
+      "implementation_from_dafny.rs"
+    );
+    Path tmpPath = null;
+    try {
+      if (Files.exists(implementationFromDafnyPath)) {
+        tmpPath = outputDir.resolve("implementation_from_dafny.rs");
         Files.move(implementationFromDafnyPath, tmpPath);
-        software.amazon.smithy.utils.IoUtils.rmdir(outputSrcDir);
-        outputSrcDir.toFile().mkdirs();
-        Files.move(tmpPath, implementationFromDafnyPath);
-      } catch (IOException e) {
-          throw new RuntimeException(e);
       }
-
+      software.amazon.smithy.utils.IoUtils.rmdir(outputSrcDir);
+      outputSrcDir.toFile().mkdirs();
+      if (tmpPath != null) {
+        Files.move(tmpPath, implementationFromDafnyPath);
+      }
+    } catch (IOException e) {
+      throw new RuntimeException(e);
+    }
 
     handlePatching(TargetLanguage.RUST, outputDir);
   }
